separate categorybutton into separate <SubCategoryButton/>
